+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/kotlin/String.kt
@@ -5,37 +5,18 @@
 package mozilla.components.support.ktx.kotlin
 
 import android.net.Uri
-import android.text.TextUtils
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
 
 /**
- * Normalizes a URL String.
- */
-fun String.toNormalizedUrl(): String {
-    val trimmedInput = this.trim()
-    var uri = Uri.parse(trimmedInput)
-    if (TextUtils.isEmpty(uri.scheme)) {
-        uri = Uri.parse("http://$trimmedInput")
-    }
-    return uri.toString()
-}
-
-/**
  * A collection of regular expressions used in the `is*` methods below.
  */
 private val re = object {
-    val urlish = "^\\s*\\w+(://|:|\\.)\\w+\\S*\\s*$".toRegex()
     val phoneish = "^\\s*tel:\\S?\\d+\\S*\\s*$".toRegex(RegexOption.IGNORE_CASE)
     val emailish = "^\\s*mailto:\\w+\\S*\\s*$".toRegex(RegexOption.IGNORE_CASE)
     val geoish = "^\\s*geo:\\S*\\d+\\S*\\s*$".toRegex(RegexOption.IGNORE_CASE)
 }
-
-/**
- * Checks if this String is a URL.
- */
-fun String.isUrl() = re.urlish.matches(this)
 
 fun String.isPhone() = re.phoneish.matches(this)